Move construct_chat_history method back to conversation.utils.py
